Matches up swaggerhub children structure to testing children structure and adds scraper
This commit allows the parser to be able to create fully form testing config specifications. Also, it adds different options for the location of the swaggerhub spec. It can support directly getting a json, getting the name of a file where the JSON is, or getting the main URL for the JSON where it can then go and download it